% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_keys}
\alias{get_keys}
\title{Get keys from a JSON input}
\usage{
get_keys(js)
}
\arguments{
\item{js}{length 1 character vector (the JSON string)}
}
\description{
More info to hack at \url{http://zserge.com/jsmn.html}
}
\examples{
get_keys('{ "dave" : "me", "henry" : 4, "steve" : [ "a", "b", "c" ]}')
}

